feat: moves eventHandlers
